renamed AnalyticsCounter -> MainClass
