--- Project02Eclipse/src/com/hemebiotech/analytics/AnalyticsCounter.java
@@ -1,17 +1,0 @@
-package com.hemebiotech.analytics;
-
-import java.util.List;
-import java.util.TreeMap;
-import java.util.TreeSet;
-
-
-public class AnalyticsCounter {
-
-	public static void main(String... args) {
-		ISymptomAnalyser analytics = new AnalyseSymptoms();
-		List<String> symptoms = analytics.getSymptoms();
-		TreeSet<String> sortedSymptoms = analytics.sortSymptoms(symptoms);
-		TreeMap<String,Integer> mappedSymptoms = analytics.countSymptoms(symptoms, sortedSymptoms);
-		analytics.writeSymptom(mappedSymptoms);
-	}
-}
